requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
sentence-transformers
streamlit_chat
openai
python-dotenv
diffusers==0.10.2
transformers 
scipy 
ftfy 
accelerate